+++ Samples/PrisonYard/Scripts/Editor/PrisonYardPaths.cs
@@ -32,7 +32,7 @@ namespace Project.Samples.PrisonYard.Editor
         public static string OutputRoot => BehaviorLLM.Editor.SampleAssetPaths.OutputRoot(Root, "PrisonYard");
         public static string Scenes => OutputRoot + "/Scenes";
         public static string Data => OutputRoot + "/Data";
-        private static string Art => BehaviorLLM.Editor.SampleAssetPaths.PrepareArt(Root, "PrisonYard");
+        private static string Art => OutputRoot + "/Art";
         public static string CharacterArt => Art + "/Characters";
         public static string DungeonArt => Art + "/Dungeon";